// src/app/interfaces/personal.interface.ts
export interface ContactInfo {
  phoneNumber: string;
  email: string;
  address: string;
  website: string;
}

export interface PersonalInfo {
  summary: string;
  firstName: string;
  lastName: string;
  dob: string;
  nationality: string;
  phoneNumber: string;
  email: string;
  address: string;
  languages: string;
}

export interface FunFacts {
  name: string;
  value: number;
}